+++ packages/create-hydrogen-app/CHANGELOG.md
@@ -7,7 +7,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## Unreleased -->
 
-- fix: add check for products.length in `Welcom.server.jsx`
+- fix: add check for products.length in `Welcome.server.jsx`
 
 ## 0.6.4 - 2021-11-11