+++ packages/create-hydrogen-app/CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+- Devcontainer support added [#164](https://github.com/Shopify/hydrogen/pull/164)
 
 ## 0.6.1 - 2021-11-08